requirements.txt
ollama
black
PyYAML
rich
